chore(core): move isNxExecutor out of workspaces
