+++ packages/cypress/src/executors/cypress/cypress.impl.ts
@@ -5,11 +5,11 @@ import {
   readTargetOptions,
   runExecutor,
   stripIndents,
-  Workspaces,
   Target,
   targetToTargetString,
   output,
 } from '@nx/devkit';
+import { getExecutorInformation } from 'nx/src/command-line/run/executor-utils';
 import 'dotenv/config';
 import { existsSync, readdirSync, unlinkSync, writeFileSync } from 'fs';
 import { basename, dirname, join } from 'path';
@@ -421,9 +421,12 @@ ${e.message || e}`);
       context.projectsConfigurations?.projects?.[target.project];
     const targetConfig = projectConfig.targets[target.target];
 
-    const workspace = new Workspaces(context.root);
     const [collection, executor] = targetConfig.executor.split(':');
-    const { schema } = workspace.readExecutor(collection, executor);
+    const { schema } = getExecutorInformation(
+      collection,
+      executor,
+      context.root
+    );
 
     // NOTE: schema won't have a default since readTargetOptions would have
     // already set that and this check wouldn't need to be made
