+++ packages/nx/src/tasks-runner/tasks-schedule.spec.ts
@@ -3,6 +3,7 @@ import { Workspaces } from '../config/workspaces';
 import { removeTasksFromTaskGraph } from './utils';
 import { Task, TaskGraph } from '../config/task-graph';
 import { DependencyType, ProjectGraph } from '../config/project-graph';
+import * as executorUtils from '../command-line/run/executor-utils';
 
 function createMockTask(id: string): Task {
   const [project, target] = id.split(':');
@@ -43,20 +44,20 @@ describe('TasksSchedule', () => {
         roots: ['lib1:build', 'app2:build'],
       };
       const workspace: Partial<Workspaces> = {
-        readExecutor() {
-          return {
-            schema: {
-              version: 2,
-              properties: {},
-            },
-            implementationFactory: jest.fn(),
-            batchImplementationFactory: jest.fn(),
-          } as any;
-        },
         readNxJson() {
           return {};
         },
       };
+      jest.spyOn(executorUtils, 'getExecutorInformation').mockReturnValue({
+        schema: {
+          version: 2,
+          properties: {},
+        },
+        implementationFactory: jest.fn(),
+        batchImplementationFactory: jest.fn(),
+        isNgCompat: true,
+        isNxExecutor: true,
+      });
 
       const projectGraph: ProjectGraph = {
         nodes: {
@@ -269,20 +270,20 @@ describe('TasksSchedule', () => {
         roots: ['app1:test', 'app2:test', 'lib1:test'],
       };
       const workspace: Partial<Workspaces> = {
-        readExecutor() {
-          return {
-            schema: {
-              version: 2,
-              properties: {},
-            },
-            implementationFactory: jest.fn(),
-            batchImplementationFactory: jest.fn(),
-          } as any;
-        },
         readNxJson() {
           return {};
         },
       };
+      jest.spyOn(executorUtils, 'getExecutorInformation').mockReturnValue({
+        schema: {
+          version: 2,
+          properties: {},
+        },
+        implementationFactory: jest.fn(),
+        batchImplementationFactory: jest.fn(),
+        isNgCompat: true,
+        isNxExecutor: true,
+      });
 
       const projectGraph: ProjectGraph = {
         nodes: {
